services/alpha-engine/requirements.txt

# Alpha Engine Dependencies
uvloop>=0.19.0
redis>=5.0.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
orjson>=3.9.0  # Fast JSON parsing
loguru>=0.7.0  # Better logging
google-cloud-pubsub>=2.18.0
google-generativeai>=0.4.0
